findmarket - scrap/requirements.txt
arrow==1.2.3
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
MouseInfo==0.1.3
mysql-connector==2.2.9
outcome==1.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
Unidecode==1.3.6
urllib3==1.26.12
wsproto==1.2.0
